# swagger/static/swagger.yaml
openapi: 3.0.0
info:
  title: API Gateway Proxy
  description: Swagger documentation for the API Gateway proxy paths.
  version: 1.0.0
servers:
  - url: http://ingress-nginx-controller.ingress-nginx.svc.cluster.local:80
paths:
...
  /users:
    get:
      summary: Get all users
      description: Proxy endpoint to fetch all users.
      responses:
        '200':
          description: List of users.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    username:
                      type: string
                    email:
                      type: string
                    role:
                      type: string
        '500':
          description: Internal server error.
  /edit-user:
    put:
      summary: Edit user details
      description: Proxy endpoint to edit user details.
      requestBody:
        description: User details to update.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                username:
                  type: string
                email:
                  type: string
                role:
                  type: string
      responses:
        '200':
          description: User details updated successfully.
        '400':
          description: Bad request.
        '500':
          description: Internal server error.
  /delete-user:
    delete:
      summary: Delete a user
      description: Proxy endpoint to delete a user.
      parameters:
        - name: id
          in: query
          required: true
          description: ID of the user to delete.
          schema:
            type: string
      responses:
        '200':
          description: User deleted successfully.
        '400':
          description: Bad request.
        '500':
          description: Internal server error.
...
  /devices/{id}:
    get:
      summary: Get device details
      description: Proxy endpoint to fetch details of a specific device.
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the device.
          schema:
            type: string
      responses:
        '200':
          description: Device details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  status:
                    type: string
                  consumption:
                    type: string
                  user_id:
                    type: string
        '404':
          description: Device not found.
        '500':
          description: Internal server error.
...